Support timeout==0 in esp_vfs_select.
The specification of select() in POSIX says that if timeout is {0, 0} then
select() shall not block the calling thread. This API call is intended
to be used for polling file descriptors. The prior implementation of esp32's
select however took a 1 tick long sleep in this situation.

The correct behavior is to not wait for the semaphore at all, but return
immediately with whatever fd we found to be active.